Fix warning on x86_64.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@38758 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ E16/e/src/lang.c
@@ -389,7 +389,7 @@ LangInit(void)
      {
 	Eprintf("Locale: %s\n", setlocale(LC_ALL, NULL));
 	Eprintf("Character encoding: locale=%s internal=%s MB_CUR_MAX=%d\n",
-		enc_loc, enc_int, MB_CUR_MAX);
+		enc_loc, enc_int, (int)MB_CUR_MAX);
      }
 
    if (!Estrcasecmp(enc_loc, "utf8") || !Estrcasecmp(enc_loc, "utf-8"))
